+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file. Guppy adher
 
 ---
 
+#### Fixed
+
+* Resolved Swift 5.1 warning on Xcode 11.2
+
+---
+
 ## Releases
 
 ---