+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file. Guppy adher
 
 * Resolved Swift 5.1 warning on Xcode 11.2
 
+#### Added
+
+* Supporting dark mode on iOS 13
+
 ---
 
 ## Releases