+++ CHANGELOG.md
@@ -11,6 +11,10 @@ All notable changes to this project will be documented in this file. Guppy adher
 * Resolved Swift 5.1 warning on Xcode 11.2
 * Fixed search bar appearance with iOS 13
 
+#### Added
+
+* Supporting dark mode on iOS 13
+
 ---
 
 ## Releases